refactor(play-screens): use game hook instead of direct store subscription (#289)
Three play-screen components were duplicating logic already computed by
their game hooks:

- WordScramblePlayScreen: replace useWordScrambleStore(useShallow(s=>s))
  with useWordScrambleGame()
- MathRacePlayScreen: replace useMathRaceStore + manual accuracy calc
  with useMathRaceGame() which already exposes accuracy
- TrueFalsePlayScreen: replace useTrueFalseStore + manual q=deck[idx]
  with useTrueFalseGame() which already exposes q

Each change removes the direct store import, the useShallow import, and
any duplicated derived-value computation.
