+++ gamesformykids/app/games/math-race/components/MathRacePlayScreen.tsx
@@ -1,13 +1,10 @@
 'use client';
-import { useShallow } from 'zustand/react/shallow';
-import { useMathRaceStore, GAME_TIME } from '../mathRaceStore';
+import { useMathRaceGame, GAME_TIME } from '../useMathRaceGame';
 import MathRaceProgressBar from './MathRaceProgressBar';
 import MathRaceQuestion from './MathRaceQuestion';
 
 export default function MathRacePlayScreen() {
-  const { q, score, timeLeft, feedback, streak, correct, total, tap } =
-    useMathRaceStore(useShallow((s) => s));
-  const accuracy = total > 0 ? Math.round((correct / total) * 100) : 0;
+  const { q, score, timeLeft, feedback, streak, correct, total, tap, accuracy } = useMathRaceGame();
 
   return (
     <div className="min-h-screen bg-gradient-to-br from-blue-100 to-indigo-200 flex flex-col items-center justify-center p-4 select-none" dir="rtl">
